test(monitoramento): não alterar a configuração de monitoramento no teste de mutação
